Withdraw duplicated function
As pointed for @fjuniorr  is_datapackage_present function was duplicated in functions and validations modules.
At first, is_datapackage_present was created in the functions module, but after a while, we realize that some validations must be realized before datasets publications/updates.
To better code organization, the validations module was created and function migrated, but is_datapackage_present stays in the functions module wrongly.

Resolves:  First observation of https://github.com/dados-mg/dpckan/pull/34#issuecomment-872613792