[#38] Correct imports in tests files
One of the main objectives of dpckan 2º iteration (#15) is the code refactor process.
Because of "readability counts" we're trying to follow the PEP 8 style guide for python code or the small_snack_case patter.
Function lerDadosJsonMapeado no longer exists so its importations in tests files must be removed
Function load_complete_datapackage is imported but needs some tests yet.

Resolves: Chunks of #38