[#15] Refactore resource_create function
One of the main objectives of dpckan 2º iteration is the code refactor process.
The original package uses CKAN API library (https://docs.ckan.org/en/2.9/api/#action-api-reference). We've discover that ckanapi library (https://github.com/ckan/ckanapi) is best for the work because It is easier and cleaner than the first library.

Resolves: Chunks of #15